Mouse Synthesizer
This page represents a simple synthesizer who pitch can be manipulated by moving
the mouse over an element -- the pitch_box in our case.

The page is comprised of three parts: the elements, its styling and the
synthesizer script. The html page and the style sheet are pretty much
self-explanatory. All that really needs to be said is that care should be taken
in naming the elements. The class names and id names should be clear and
readable. Also, the script uses the names to load the elements as objects, so it
helps to have clear indication of what each element represents.

The synthesizer script creates sound using the browser's WebAudio API and it
listens for DOM events on two elements in the page.

In the script, a reference to the WebAudio API is created and is used to create
the oscillator, connect the oscillator to the audio output, and start the
oscillator. A listener onmousemove also changes the frequency value of the
oscillator, when the mouse moves over the pitch_box, changing the oscillators'
pitch. We also use the click event to stop the oscillator.

One last note: use window.onload to start the script rather than invoking it
immediately. If you invoke it immediately, note that it will start when its
initially loaded. So if you place the script tag before the elements of the
page, then it will try to access elements of the page before the DOM has even
been formed. You can either move the script tag to the end of the page or you
can just use window.onload.

mouse_synth.html:
        - HTML for the page.
                - Started with the word STOP, but ended up using the square
                  symbol for stop -- you know, the ones you used to see on VCRs
                  and tape decks.

style.css:
        - I borrowed the material design drop shadow guidelines to give the page
          some pop. Thanks to the codepen:
          https://codepen.io/sdthornton/pen/wBZdXq

mouse_synth.js:
        - uses the WebAudioAPI's oscillator to create sound

        - uses event listeners to change the pitch of the oscillator when the
          mouse moves over the element 'pitch_box'

        - stops the sound when a click event occurs on the stop button.