LFO Modulator
This is a modulator that allows us to modulate the sound of an oscillator using
the mouse.

To create the modulator, we need to modulate some type of sound. So we use an
oscillator to produce this reference sound. For the initial sound, we feed into
this oscillator's frequency, a low frequency oscillator. The LFO will manipulate
the frequency of the oscillator, thus modulating the sound.

We also add mouse controls that allows us to change the lfo's frequency.
Obviously if we change the frequency of the LFO to a high value, it stops being
an LFO. But we can move to the left to decrease the "LFO"'s frequency.

Because the modulation produced by the LFO is subtle, we exaggerate the sound by
running the LFO's output through a gain node. The gain node amplifies the LFO,
then we run it into the oscillator's frequency. So, the oscillator's frequency
will change according to the LFO's output (I'm not sure what values frequency
takes on). (I'm also not sure what gain the LFO experiences, but, in
general, it means increasing the power of the signal.)

I'm still not sure it's a good idea to maintain the state of the control button
in two different scripts. Initially, it made sense to separate the audio code
from the presentation code. But, the state of the button affects both the audio
and the style of button. This prompts the following questions: Won't this cause
problems down the line since the state is being maintained in two different
contexts? Should the state be maintained in one place or is consistency across
all files enough to maintain the state? The latter for the second question seems
like it would require us to keep track of the state in all files meticulously --
which seems ripe for mistakes to occur. For a small app like this, it's easy to
keep track of the state. For larger ones, we may need to find an alternative
way. I'm still not clear on what the organization should be for this type of
situation.

controls.js:
        - Controls the style of the button. The icon changes to reflect its ON
          state and OFF state when clicked.
                - There has to be a better way to change the styling without
                  having to manipulate every single attribute. I've seen the use
                  of the toggle method in conjunction with a .toggle class, but
                  I don't recall the specifics. I have to look into it.
lfo.js:
        - We create a modulator that uses a low frequency oscillator to
          manipulate the frequency of another oscillator. But before we feed the
          lfo's output into the oscillator's frequency, we amplify the lfo by
          increasing the gain. This gives us an exaggerated sound so we can
          hear the changes in frequency, caused by the LFO, of the oscillator.

        - The change in the mouse's x position changes the lfo's frequency.
lfo.html:
        - The LFO modulator page
style.css:
        - The styling for the page
        - The page uses height: 100% for the modulator element, but the property
          can't exist alone. The property uses a percentage which inherits its
          height from body. It's not enough to set the body's height to 100%
          because its height is set to 0 for the html element. Therefore, both
          the html's and the body's height have to be set to 100%.