Refactoring the LFO_Modulation Project
In the previous version of this program, I had two separate files: one was
tasked with maintaining the audio and another was tasked with updating the style
of the controls. Seems reasonable right? The problem was that the on/off state
that determines the result of the tasks couldn't be shared across two different
files unless we created a global variable for the on/off state. I wanted to
avoid doing this, I instead maintained two isolated states in each file with the
idea that in the abstract they we're the same. As long as the manipulations of
the state were kept consistent, every thing would be fine. Turns out I failed
because not even I could keep the state consistent across both files. It
extremely easy to make an error by keeping track of the state in two different
files. No you can understand why I refactored.

The refactor isolates the setup and "exports" an interface to manipulate the
audio and the style.  The state is kept in one context, instead of two, so there
won't be any problems keeping track of it. One of the reasons that an interface
is returned is because at times we need to alter a variable in a different
scope (functions create scopes that contain references initialized with var).
Seperating the setups into two makes writing the event handlers in the onload
callback relatively effortless.

I'll continue to refactor this as time goes on. One idea I have is to remove the
manual styling I have in setupControls and replace by just changing the class
name of the element. The browser will have to apply the appropriate style for
that class when the change occurs.

/e_music/LFO_Modulation
control.js:
        - Refactored the script so that the state is maintained in one place
          instead of two places. The state wasn't even consistent across the two
          files.

        - The setup routines return an object containing methods necessary
          to update the audio and the styles based on the state.

lfo.html:
        - Removing the controls script now that the controls are handled by
          lfo.js